requirements.txt
requests
supabase
python-dotenv
rapidfuzz
